requirements.txt
fastapi
sqlalchemy
mysql-connector-python  
aiomysql
alembic
uvicorn
passlib
pyjwt
python-dotenv